requirements.txt
Package                   Version
------------------------- -----------
attrs                     25.4.0
certifi                   2025.10.5
charset-normalizer        3.4.3
colorama                  0.4.6
deepdiff                  8.0.1
et_xmlfile                2.0.0
execnet                   2.1.1
idna                      3.10
iniconfig                 2.1.0
Jinja2                    3.1.6
jsonschema                4.23.0
jsonschema-specifications 2025.9.1
MarkupSafe                3.0.3
numpy                     2.3.3
openpyxl                  3.1.5
orderly-set               5.2.2
packaging                 25.0
pandas                    2.2.3
pip                       23.2
pluggy                    1.6.0
pytest                    8.3.3
pytest-html               4.1.1
pytest-json-report        1.5.0
pytest-metadata           3.1.1
pytest-order              1.3.0
pytest-xdist              3.6.1
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
pytz                      2025.2
referencing               0.36.2
requests                  2.32.3
rpds-py                   0.27.1
setuptools                68.1.2
six                       1.17.0
tqdm                      4.66.5
typing_extensions         4.15.0
tzdata                    2025.2
urllib3                   2.5.0
xmltodict                 0.13.0
